@@ -1,1 +1,1 @@
-2.5.0
+2.6.0
