@@ -58,8 +58,8 @@ describe Doorkeeper::OAuth::IdTokenRequest do
 
     it 'creates a new token if scopes do not match' do
       allow(Doorkeeper.configuration).to receive(:reuse_access_token).and_return(true)
-      FactoryGirl.create(:access_token, application_id: pre_auth.client.id,
-                         resource_owner_id: owner.id, scopes: '')
+      create(:access_token, application_id: pre_auth.client.id,
+             resource_owner_id: owner.id, scopes: '')
       expect do
         subject.authorize
       end.to change { Doorkeeper::AccessToken.count }.by(1)
@@ -70,8 +70,8 @@ describe Doorkeeper::OAuth::IdTokenRequest do
       allow(application.scopes).to receive(:has_scopes?).and_return(true)
       allow(application.scopes).to receive(:all?).and_return(true)
 
-      FactoryGirl.create(:access_token, application_id: pre_auth.client.id,
-                         resource_owner_id: owner.id, scopes: 'public')
+      create(:access_token, application_id: pre_auth.client.id,
+             resource_owner_id: owner.id, scopes: 'public')
 
       expect { subject.authorize }.not_to change { Doorkeeper::AccessToken.count }
     end
